Changed max_keywords to be of type size_t

+++ pappl/lookup.c
@@ -45,7 +45,7 @@ _papplLookupString(
 size_t        // O - number of keywords returned
 _papplLookupStrings(
     unsigned            value,        // I - Bitfield value
-    int                 max_keywords, // I - Max number of keywords to return
+    size_t                 max_keywords, // I - Max number of keywords to return
     char          *keywords[],  // I - Array of char * to be assigned and retuned
     size_t              num_strings,  // I - Number of strings
     const char * const  *strings)     // I - Strings
@@ -54,7 +54,7 @@ _papplLookupStrings(
   unsigned  bit;            // Current bit
   size_t    counter = 0;    // Counter to be returned
 
-  for (i = 0, bit = 1; i < num_strings && counter < (size_t)max_keywords; i ++, bit *= 2)
+  for (i = 0, bit = 1; i < num_strings && counter < max_keywords; i ++, bit *= 2)
   {
     if (value & bit)
     {
